refactor(niivue): share volume render data helpers
Move RGBA datatype detection and PAQD overlay preparation into a shared helper used by both WebGL2 and WebGPU renderers.
